requirements.txt
Flask==3.1.0
numpy==1.24.3
Pillow==11.1.0
Requests==2.32.3
sentence_transformers==3.2.1
tqdm==4.66.5
Werkzeug==3.1.3
